+++ application/packages/types/src/core/index.ts
@@ -1,2 +1,3 @@
 export * from './capture/types.capture';
 export * from './types/types.types';
+export * from './replay/replay';